requirements.txt
Flask == 3.1.2
opencv-python == 4.12.0.88
RapidFuzz == 3.14.1
Flask-SQLAlchemy == 3.1.1
Flask-Login ==  0.6.3
PyMySQL == 1.1.2
python-dotenv == 1.1.1
openai  == 1.108.1
google-cloud-documentai == 3.6.0
google-cloud-vision == 3.10.2

#ApiKey para GPT:
#URL:https://platform.openai.com/api-keys

#ApiKey para ORC Cloud Vision:
#URL:https://console.cloud.google.com/apis/credentials